requirements.txt
Django~=3.2.5
django-ckeditor==6.1.0